feat: use scan qr button from the home for reward wallet
