+++ src/app/(mobile-ui)/wallet/page.tsx
@@ -6,7 +6,6 @@ import DirectionalActionButtons from '@/components/Global/DirectionalActionButto
 import NoDataEmptyState from '@/components/Global/EmptyStates/NoDataEmptyState'
 import Icon from '@/components/Global/Icon'
 import { ListItemView } from '@/components/Global/ListItemView'
-import QRScanner from '@/components/Global/QRScanner'
 import { PartnerBarLocation, RewardDetails } from '@/components/Global/RewardsModal'
 import { WalletCard } from '@/components/Home/WalletCard'
 import { PEANUT_WALLET_CHAIN, PEANUT_WALLET_TOKEN_DECIMALS, PEANUT_WALLET_TOKEN_NAME } from '@/constants'
@@ -19,17 +18,14 @@ import {
     formatAmount,
     getChainName,
     getUserPreferences,
-    resolveFromEnsName,
     updateUserPreferences,
-    validateEnsName,
 } from '@/utils'
 import { useDisconnect } from '@reown/appkit/react'
-import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-import { formatUnits, isAddress } from 'viem'
+import { formatUnits } from 'viem'
+import DirectSendQr from '@/components/Global/DirectSendQR'
 
 const WalletDetailsPage = () => {
-    const router = useRouter()
     const { disconnect } = useDisconnect()
     const { focusedWallet: focusedWalletId } = useWalletStore()
     const { connectWallet } = useWalletConnection()
@@ -39,39 +35,6 @@ const WalletDetailsPage = () => {
         const prefs = getUserPreferences()
         return prefs?.balanceHidden ?? false
     })
-    const [isQRScannerOpen, setIsQRScannerOpen] = useState(false)
-
-    // Process QR code data
-    const processQRCode = async (data: string) => {
-        const baseUrl = process.env.NEXT_PUBLIC_BASE_URL!
-        if (data.startsWith(baseUrl)) {
-            let path = data
-            if (path.startsWith(baseUrl)) {
-                path = path.substring(baseUrl.length)
-            }
-            if (!path.startsWith('/')) {
-                path = '/' + path
-            }
-            router.push(path)
-            return { success: true }
-        } else if (isAddress(data)) {
-            // hardcode PNT because this is only used for pinta wallet for now
-            const link = `${baseUrl}/${data}@polygon/PNT`
-            router.push(link)
-            return { success: true }
-        } else if (validateEnsName(data)) {
-            const resolvedAddress = await resolveFromEnsName(data.toLowerCase())
-            if (!!resolvedAddress) {
-                const link = `${baseUrl}/${resolvedAddress}@polygon/PNT`
-                router.push(link)
-                return { success: true }
-            }
-        }
-        return {
-            success: false,
-            error: 'QR not recognized as Peanut URL',
-        }
-    }
 
     const walletDetails = wallets.find((wallet) => wallet.id === focusedWalletId)
     const isPeanutWallet = walletDetails?.walletProviderType === WalletProviderType.PEANUT
@@ -207,20 +170,7 @@ const WalletDetailsPage = () => {
                 </>
             )}
 
-            {isRewardsWallet && (
-                <Button
-                    onClick={() => setIsQRScannerOpen(true)}
-                    variant="purple"
-                    className={'mx-auto flex w-fit cursor-pointer items-center justify-center gap-2 rounded-full'}
-                    shadowSize="4"
-                >
-                    <Icon name="camera" fill="black" className="size-4" />
-                    <span>Scan QR Code</span>
-                </Button>
-            )}
-            {isQRScannerOpen && (
-                <QRScanner onScan={processQRCode} onClose={() => setIsQRScannerOpen(false)} isOpen={true} />
-            )}
+            {isRewardsWallet && <DirectSendQr />}
 
             {renderRewardDetails()}
         </div>
